functional record and loop feature. still needs more work USE AT OWN RISK
if you use this you will need to delete all the recordings in order to reset and record something new without playing the old stuff. In order to do this you need to uncomment lines in the playbackHandler constructor. the stop button also works like a play button if you havent hit start recording.